Refactor breadcrumbs to work with publishing api
The code for flattening nested links hashes into breadcrumbs is useful in other
GOV.UK apps besides navigation, for example:

- in rummager we want to tag a document with all its breadcrumb topics at
  indexing time
- in whitehall we want to show a preview of the breadcrumb for every topic we
  tag to

The existing Breadcrumb and TaxonBreadcrumb classes worked on ContentItem
objects, and this class parses content store representations of content items.

In publishing apps, we call a publishing api method, `GET expanded-links`,
which has a slightly different response, but with the same basic structure for
links.

We've added a new class to parse expanded-links responses, and made
TaxonBreadcrumb and Breadcrumb work with both of them.

This changes the API for using these classes, but keeps NavigationHelper the
same.