Use document supertype to fetch guidance content
Use the `filter_navigation_document_supertype` field which has recently
been added to search when fetching guidance content to display in the
navigation pages.

This new field encapsulates the same logic and is being populated during
indexing in rummager and will be always present.

Also delete the hard-coded list of guidance document types, because the
only app which refered to this directly (collections) has also switched
to using the search field.