Whitelist root taxons
Since we’ve created the new single taxonomy in draft in Production, some publishers have been seeing some of those draft taxons in the sidebar and breadcrumbs when previewing their content on the draft stack.

In order to avoid this, we add a hardcoded whitelist of taxons which we want to show (including all their sub taxons). This whitelist corresponds to the current published branches World, Education and Parenting. We currently don’t expect to be publishing any new taxonomy branches, so we think that this approach will work for the time being. We expect to remove this whitelist when the single taxonomy is more settled and/or is published.