Fix failing taxon breadcrumb spec
This test was failing on master. It seems to be looking for the
'guidance' schema which doesn't exist (anymore?). Instead we update
the spec to use the 'answer' document type.

In the `breadcrumbs_for` method we should be fetching an example
for the given content item's document type, not 'taxon' as was
previously the case (which was another reason for failing the spec).

Update the context wording around 'taxon with taxon parents' and
'taxon content item with taxon parents’ as I believe those two were
the wrong way around.

We also update the spec to use 2 space indentation.