Update taxonomy sidebar to contain collection links
We have removed the related navigation helper from this gem, however the part
of `government-frontend` that constructs links for the taxonomy sidebar relies
on data related to collections which is provided by this helper. As we are
already constructing link data for the taxonomy sidebar in the
`TaxonomySidebar` nav helper, we add collections as a type of link that it
should construct as well.